Cleaning the creation of the director
