+++ circusort/base/director.py
@@ -4,7 +4,26 @@ import logging
 from circusort.base.logger import Logger
 from circusort.base.utils import get_log
 from circusort.base.process import create_process
+from circusort.base.utils import find_interface_address_towards
 
+
+def create_director(host='127.0.0.1', **kwargs):
+    """Create a new director in this process.
+
+    Parameter:
+        host: string (optional)
+            The IP address of the host of the director.
+    Return:
+        director: circusort.base.director
+            The director.
+    See also:
+        circusort.base.director.Director
+    """
+
+    interface = find_interface_address_towards(host)
+    director = Director(interface, **kwargs)
+
+    return director
 
 
 class Director(object):
@@ -25,19 +44,23 @@ class Director(object):
         self.managers = {}
 
     def __del__(self):
+
         self.log.info("{d} is destroyed".format(d=str(self)))
+
         for manager in self.managers.itervalues():
             manager.__del__()
 
     @property
     def nb_managers(self):
+
         return len(self.managers)
 
     def get_logger(self):
+
         return self.logger
 
     def create_block(self, block_type, name=None, log_level=None, **kwargs):
-        '''TODO add docstring'''
+        # TODO add docstring.
 
         if self.nb_managers == 0:
             self.create_manager(log_level=self.log_level)
@@ -47,10 +70,11 @@ class Director(object):
         return block
 
     def create_manager(self, name=None, host=None, log_level=None):
-        '''Create a new manager process and return a proxy to this process.
+        """Create a new manager process and return a proxy to this process.
 
         A manager is a process that manages workers.
-        '''
+        """
+
         if name is None:
             manager_id = 1 + self.nb_managers
             name = "Manager {}".format(manager_id)
@@ -71,10 +95,11 @@ class Director(object):
         
         self.managers.update({manager.name: manager})
         self.log.debug("{d} registers {m}".format(d=str(self), m=manager.name))
+
         return
 
     def connect(self, output_endpoints, input_endpoints, protocol=None):
-        '''TODO add docstring'''
+        # TODO add docstring.
 
         if not type(input_endpoints) == list:
             input_endpoints = [input_endpoints]
@@ -92,7 +117,8 @@ class Director(object):
         for input_endpoint in input_endpoints:
             for output_endpoint in output_endpoints:
 
-                self.log.info("{d} connects {s} to {t}".format(d=str(self), s=output_endpoint.block.name, t=input_endpoint.block.name))
+                self.log.info("{d} connects {s} to {t}".format(d=str(self), s=output_endpoint.block.name,
+                                                               t=input_endpoint.block.name))
 
                 if input_endpoint.block.parent == output_endpoint.block.parent:
                     if protocol is None:
@@ -102,7 +128,8 @@ class Director(object):
                             local_protocol = 'tcp'
                     else:
                         local_protocol = protocol
-                    self.get_manager(input_endpoint.block.parent).connect(output_endpoint, input_endpoint, local_protocol, show_log=False)
+                    self.get_manager(input_endpoint.block.parent).connect(output_endpoint, input_endpoint,
+                                                                          local_protocol, show_log=False)
                 else:
                     if protocol is None:
                         local_protocol = 'tcp'
@@ -115,56 +142,79 @@ class Director(object):
                     input_endpoint.configure(**description)
                     input_endpoint.block.connect(input_endpoint.name)
                     input_endpoint.block.guess_output_endpoints()
-                    self.log.debug("{p} connection established from {a}[{s}] to {b}[{t}]".format(p=local_protocol, s=(output_endpoint.name, output_endpoint.structure), 
-                                                                                                    t=(input_endpoint.name, input_endpoint.structure), 
-                                                                                                    a=output_endpoint.block.name,
-                                                                                                    b=input_endpoint.block.name))
-             
+
+                    string = "{p} connection established from {a}[{s}] to {b}[{t}]"
+                    message = string.format(p=local_protocol, s=(output_endpoint.name, output_endpoint.structure),
+                                            t=(input_endpoint.name, input_endpoint.structure),
+                                            a=output_endpoint.block.name, b=input_endpoint.block.name)
+                    self.log.debug(message)
+
         return
 
     def initialize(self):
+
         self.log.info("{d} initializes {s}".format(d=str(self), s=", ".join(self.list_managers())))
         for manager in self.managers.itervalues():
             manager.initialize()
+
         return
 
     def start(self, nb_steps=None):
+
         if nb_steps is None:
-            self.log.info("{d} starts {s}".format(d=str(self), s=", ".join(self.list_managers())))
+            string = "{d} starts {s}"
+            message = string.format(d=str(self), s=", ".join(self.list_managers()))
+            self.log.info(message)
         else:
-            self.log.info("{d} runs {s} for {n} steps".format(d=str(self), s=", ".join(self.list_managers()), n=nb_steps))
+            string = "{d} runs {s} for {n} steps"
+            message = string.format(d=str(self), s=", ".join(self.list_managers()), n=nb_steps)
+            self.log.info(message)
+
         for manager in self.managers.itervalues():
             manager.start(nb_steps)
+
         return
 
     def sleep(self, duration=None):
+
         self.log.info("{d} sleeps for {k} sec".format(d=str(self), k=duration))
         time.sleep(duration)
+
         return
 
     def stop(self):
+
         self.log.debug("{d} stops {s}".format(d=str(self), s=", ".join(self.list_managers())))
         for manager in self.managers.itervalues():
             manager.stop()
         self.log.info("{d} stops {s}".format(d=str(self), s=", ".join(self.list_managers())))
+
         return
 
     def join(self):
+
         self.log.debug("{d} joins {s}".format(d=str(self), s=", ".join(self.list_managers())))
         for manager in self.managers.itervalues():
             manager.join()
         self.log.info("{d} joins {s}".format(d=str(self), s=", ".join(self.list_managers())))
+
         return
 
     def destroy(self):
+
         self.__del__()
 
+        return
+
     def __str__(self):
+
         return "{d}[{i}]".format(d=self.name, i=self.host)
 
     def list_managers(self):
+
         return self.managers.keys()
 
     def get_manager(self, key):
+
         assert key in self.list_managers(), self.log.warning("%s is not a valid manager" %key)
         return self.managers[key]
